backend/requirements.txt
fastapi==0.64.0
uvicorn==0.13.4
pydantic==1.7.3
SQLAlchemy==1.3.20
aiomysql==0.0.20
python-dotenv==0.19.2
starlette==0.13.6
PyMySQL==0.9.2
click==7.1.2
h11==0.9.0
asgiref==3.3.4
